Leave institution creation
Not sure how it will work.
Need to authenticate that the person starting an institution represents
the school they say using. May need to be human verification - or get
some sort of reference that identifies schools.
Also cases where institution is not a school and is purely defined as
this online entity.